hamal-9 fixes
