+++ platform/lib/domain/src/main/kotlin/io/hamal/lib/domain/vo/CodeType.kt
@@ -1,0 +1,15 @@
+package io.hamal.lib.domain.vo
+
+enum class CodeType(val value: Int) {
+    None(0),
+    Lua54(1);
+
+    companion object {
+        @JvmStatic
+        fun of(value: Int): CodeType {
+            val result = CodeType.values().find { it.value == value }
+            require(result != null) { "$value not mapped as a code type" }
+            return result
+        }
+    }
+}
