+++ platform/lib/sdk/src/main/kotlin/io/hamal/lib/sdk/api/Code.kt
@@ -21,17 +21,18 @@ interface ApiCodeService {
 internal class ApiCodeServiceImpl(
     private val template: HttpTemplate
 ) : ApiCodeService {
-    override fun get(codeId: CodeId, codeVersion: CodeVersion){
-            template.get("/v1/code/{codeId}")
-                             .path("codeId", codeId)
-                             .let{builder ->
-                                      if(codeVersion != null){
-                                         builder.parameter("code_version", codeVersion.value)
-                                      }else{
-                                         builder
-                                      }
-                             }
-                                 .execute()
+
+    override fun get(codeId: CodeId, codeVersion: CodeVersion?): ApiCode {
+        return template.get("/v1/code/{codeId}")
+            .path("codeId", codeId)
+            .let { builder ->
+                if (codeVersion != null) {
+                    builder.parameter("code_version", codeVersion.value)
+                } else {
+                    builder
+                }
+            }
+            .execute()
             .fold(ApiCode::class)
     }
 }
