+++ platform/runner-extension/std/sys/src/main/kotlin/io/hamal/extension/std/sys/code/CodeGetFunction.kt
@@ -2,44 +2,15 @@ package io.hamal.extension.std.sys.code
 
 import io.hamal.lib.domain.vo.CodeId
 import io.hamal.lib.domain.vo.CodeVersion
-import io.hamal.lib.kua.function.*
+import io.hamal.lib.kua.function.Function2In2Out
+import io.hamal.lib.kua.function.FunctionContext
+import io.hamal.lib.kua.function.FunctionInput2Schema
+import io.hamal.lib.kua.function.FunctionOutput2Schema
 import io.hamal.lib.kua.type.*
 import io.hamal.lib.sdk.ApiSdk
 
 
-class CodeGetFunctionI(
-    private val sdk: ApiSdk
-) : Function1In2Out<StringType, ErrorType, MapType>(
-    FunctionInput1Schema(StringType::class),
-    FunctionOutput2Schema(ErrorType::class, MapType::class)
-) {
-
-    override fun invoke(ctx: FunctionContext, arg1: StringType, arg2: NumberType): Pair<ErrorType?, MapType?> {
-        return try {
-        
-        val response = if(args == NumberType(-1)){
-                     sdk.code.get(CodeId(arg1.value))
-        }else{
-                     sdk.code.get(CodeId(arg1.value), CodeVersion(arg2.value))
-        }
-        
-        
-            null to response .let { code ->
-                    MapType(
-                        mutableMapOf(
-                            "id" to StringType(code.id.value.value.toString(16)),
-                            "code" to CodeType(code.value.value),
-                            "version" to NumberType(code.version.value)
-                        )
-                    )
-                }
-        } catch (t: Throwable) {
-            ErrorType(t.message!!) to null
-        }
-    }
-}
-
-class CodeGetFunctionII(
+class CodeGetFunction(
     private val sdk: ApiSdk
 ) : Function2In2Out<StringType, NumberType, ErrorType, MapType>(
     FunctionInput2Schema(StringType::class, NumberType::class),
@@ -47,7 +18,13 @@ class CodeGetFunctionII(
 ) {
     override fun invoke(ctx: FunctionContext, arg1: StringType, arg2: NumberType): Pair<ErrorType?, MapType?> {
         return try {
-            null to sdk.code.get(CodeId(arg1.value), CodeVersion(arg2.value.toInt()))
+            val response = if (arg2 == NumberType(-1)) {
+                sdk.code.get(CodeId(arg1.value))
+            } else {
+                sdk.code.get(CodeId(arg1.value), CodeVersion(arg2.value.toInt()))
+            }
+
+            null to response
                 .let { code ->
                     MapType(
                         mutableMapOf(
